feat(manager-legacy-tools): delete ods 17 in utils
ref: #MANAGER-19506
