+++ packages/manager-tools/manager-legacy-tools/test-utils/package.json
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "@ovh-ux/manager-core-api": "^0.17.1",
+    "@ovhcloud/ods-components": "^18.6.2",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
@@ -31,8 +32,6 @@
   },
   "peerDependencies": {
     "@ovh-ux/manager-core-api": "^0.9.0",
-    "@ovhcloud/ods-common-core": "17.x",
-    "@ovhcloud/ods-components": "17.x",
     "@tanstack/react-query": "5.x",
     "i18next": "23.x",
     "react": "18.x"
