+++ packages/manager-tools/manager-legacy-tools/test-utils/src/utils/setup-test.ts
@@ -1,0 +1,44 @@
+import { SetupServer, setupServer } from 'msw/node';
+import { toMswHandlers } from './msw';
+import { getAuthenticationMocks } from '../auth';
+import '@testing-library/jest-dom';
+import 'element-internals-polyfill';
+
+declare global {
+  // eslint-disable-next-line vars-on-top, no-var
+  var server: SetupServer;
+  // eslint-disable-next-line vars-on-top, no-var, @typescript-eslint/naming-convention
+  var __VERSION__: string;
+}
+
+const server = setupServer(
+  ...toMswHandlers([
+    ...getAuthenticationMocks({ isAuthMocked: true, region: 'EU' }),
+  ]),
+);
+
+beforeAll(() => {
+  server.listen({ onUnhandledRequest: 'warn' });
+
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
+  delete global.server;
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
+  global.__VERSION__ = null;
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
+  global.server = server;
+});
+
+afterAll(() => {
+  server.close();
+
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
+  delete global.__VERSION__;
+});
+
+afterEach(() => {
+  server.resetHandlers();
+});
