+++ packages/manager-tools/manager-legacy-tools/test-utils/src/utils/ui-test-helpers-ods18.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { OsdsIcon } from '@ovhcloud/ods-components/react';
+import { OdsIcon } from '@ovhcloud/ods-components/react';
 import {
   screen,
   waitFor,
@@ -50,7 +50,7 @@ type GetOdsButtonParams = {
   container: HTMLElement;
   label?: string;
   altLabel?: string;
-  iconName?: React.ComponentProps<typeof OsdsIcon>['name'];
+  iconName?: React.ComponentProps<typeof OdsIcon>['name'];
   disabled?: boolean;
   isLink?: boolean;
   nth?: number;
