+++ packages/manager/apps/dedicated-servers/package.json
@@ -46,6 +46,7 @@
     "@testing-library/react": "^14.2.1",
     "@vitejs/plugin-react": "^4.2.1",
     "@vitest/coverage-v8": "^2.1.9",
+    "element-internals-polyfill": "^1.3.12",
     "msw": "2.1.7",
     "typescript": "^5.1.6",
     "vite": "^6.0.7",
