+++ packages/manager/apps/hpc-vmware-managed-vcd/vitest.config.js
@@ -10,7 +10,7 @@ export default mergeConfig(
   sharedConfig,
   createConfig({
     test: {
-      setupFiles: '@ovh-ux/manager-core-test-utils/setup-file-msw-ods17.tsx',
+      setupFiles: ['@ovh-ux/manager-core-test-utils/src/utils/setup-test.ts'],
       coverage: {
         exclude: [
           ...defaultExcludedFiles,
