+++ packages/manager/apps/veeam-backup/package.json
@@ -48,7 +48,8 @@
     "@ovh-ux/manager-tests-setup": "latest",
     "@ovh-ux/manager-vite-config": "^0.11.0",
     "msw": "2.1.7",
-    "typescript": "^5.1.6"
+    "typescript": "^5.1.6",
+    "vitest": "^2.1.9"
   },
   "regions": [
     "CA",
